Don’t block when scheduling task from outside loop
There was a lack of threadsafety from trying to schedule a task from the event message’s loop (within _GPSThread) to the main loop blocked as it didn’t ever actually got to the main event loop. This allows us to schedule tasks on a given loop from another loop.