[gpubsub_publisher] explicitly report publish error
Previously error messages could be determined by the difference
between recv and sent, but I think it will be easier to understand
an explicit error metric.  Additionally the new format uses the
same "what" with different attributes, which should make it easier
to compare the two in the same graph.