[reconciler] replace additions/deletions with rrset metric with source
The metric is changed to reuse the same "what" so it is easier to
compare additions and deletions on the same graph. Additionally,
source is added by the authority and included as an attribute.

Also, move the ResourceRecordSet object from the shared dns client
to the reconciler plugin as it is the only plugin that uses it.